chore: rename cloud-> v1-cloud to enable easy rebase
