+++ setup.v1-cloud.ts
@@ -1,10 +1,10 @@
-import { registerFireproofCloudStoreProtocol } from "./src/cloud/gateway.ts";
+import { registerFireproofCloudStoreProtocol } from "./src/v1-cloud/gateway.ts";
 import { BuildURI } from "@adviser/cement";
 
 registerFireproofCloudStoreProtocol();
 
 const url = BuildURI.from("fireproof://localhost:1998")
-  .setParam("getBaseUrl", "http://127.0.0.1:9000/testbucket/fp-cloud-test")
+  // .setParam("getBaseUrl", "http://127.0.0.1:9000/testbucket/fp-cloud-test")
   .setParam("protocol", "ws")
   .setParam("getNeedsAuth", "true");
 process.env.FP_STORAGE_URL = url.toString();
