+++ tests/docker-compose.yaml
@@ -15,11 +15,11 @@ services:
     ports:
       - 1999:1999
 
-  fp-cloud:
-    image: connect-fp-cloud
+  v1-cloud:
+    image: connect-v1-cloud
     build:
       context: ..
-      dockerfile: ./tests/Dockerfile.connect-fp-cloud
+      dockerfile: ./tests/Dockerfile.connect-v1-cloud
     ports:
       - 1998:1998
 
@@ -143,13 +143,13 @@ services:
 
   wait-for-ready:
     image: curlimages/curl
-    command: "sh -c 'while [ ! -e healthy ] ; do curl -f http://netlify:8888 > netlify && curl -f http://minio:9000/minio/health/live > minio && curl -f http://ucan:8787/health > ucan && curl -f http://partykit:1999 > partykit && curl -f http://fp-cloud:1998/health > fp-cloud && echo ready > ready ; ls; sleep 5; done ; sleep 60'"
+    command: "sh -c 'while [ ! -e healthy ] ; do curl -f http://netlify:8888 > netlify && curl -f http://minio:9000/minio/health/live > minio && curl -f http://ucan:8787/health > ucan && curl -f http://partykit:1999 > partykit && curl -f http://v1-cloud:1998/health > v1-cloud && echo ready > ready ; ls; sleep 5; done ; sleep 60'"
     depends_on:
       netlify:
         condition: service_started
       partykit:
         condition: service_started
-      fp-cloud:
+      v1-cloud:
         condition: service_started
       ucan:
         condition: service_started
