+++ vitest.v1-cloud.config.ts
@@ -5,7 +5,7 @@ import tsconfigPaths from "vite-tsconfig-paths";
 export default defineConfig({
   plugins: [tsconfigPaths()],
   test: {
-    name: "cloud",
+    name: "v1-cloud",
     exclude: [
       "node_modules/@fireproof/core/tests/react/**",
       "node_modules/@fireproof/core/tests/fireproof/config.test.ts",
@@ -14,10 +14,10 @@ export default defineConfig({
       // "node_modules/@fireproof/core/tests/**/*test.?(c|m)[jt]s?(x)",
       // "node_modules/@fireproof/core/tests/**/*gateway.test.?(c|m)[jt]s?(x)",
       // "src/connector.test.ts",
-      "src/cloud/*test.?(c|m)[jt]s?(x)",
+      "src/v1-cloud/*test.?(c|m)[jt]s?(x)",
     ],
     globals: true,
-    setupFiles: "./setup.cloud.ts",
+    setupFiles: "./setup.v1-cloud.ts",
     testTimeout: 25000,
   },
 });
