+++ vitest.workspace.ts
@@ -4,7 +4,7 @@ import aws from "./vitest.aws.config.ts";
 import betterSqlite3 from "./vitest.better-sqlite3.config.ts";
 // import nodeSqlite3Wasm from "./vitest.node-sqlite3-wasm.config.ts";
 import partykit from "./vitest.partykit.config.ts";
-import cloud from "./vitest.cloud.config.ts";
+import v1Cloud from "./vitest.v1-cloud.config.ts";
 import s3 from "./vitest.s3.config.ts";
 // import connector from "./vitest.connector.config.ts";
 import netlify from "./vitest.netlify.config.ts";
@@ -21,7 +21,7 @@ export default defineWorkspace([
   aws,
   netlify,
   partykit,
-  cloud,
+  v1Cloud,
   //cf_kv
   ucan,
 ]);
